Move dynamic search params to navigation rendering
Move the hook behind the navigation module's browser-variant boundary so the follow-up can import ReactDOM.browser() without resolving that export in the RSC app worker. This preserves the existing BailoutToCSRError behavior.
